+++ R/data.R
@@ -3,6 +3,8 @@
 #' Stations' data from the Greek National Data Bank for Hydrological and
 #' Meteorological Information databases: \url{http://kyy.hydroscope.gr/},
 #' \url{http://emy.hydroscope.gr/} and \url{http://ypaat.hydroscope.gr/}
+#' The stations with IDs: have erroneous coordinates and their longitudes,
+#' latitudes have been replaced with NA values.
 #'
 #' @format A data frame:
 #' \describe{
@@ -14,7 +16,7 @@
 #'     \item{PoliticalDivision}{The station's Political Division}
 #'     \item{Owner}{The station's owner}
 #'     \item{Type}{The station's type}
-#'     \item{Domain}{The corresponding domain}
+#'     \item{Domain}{The corresponding Hydroscope's domain}
 #'     \item{Long}{The station's longitude in decimal degrees, ETRS89}
 #'     \item{Lat}{The station's latitude in decimal degrees, ETRS89}
 #'     \item{Elevation}{The station's altitude, meters above sea level}
@@ -38,7 +40,6 @@
 #'     \item{StartDate}{The starting date of time series values}
 #'     \item{EndDate}{The ending date of time series values}
 #'     \item{StationID}{The corresponding station's ID}
-#'     \item{Name}{The time series name}
-#'     \item{Remarks}{Remarks for time series from Hydroscope}
+#'     \item{Domain}{The corresponding Hydroscope's domain}
 #' }
 "timeseries"