+++ NEWS.md
@@ -1,10 +1,11 @@
 # hydroscoper 0.1.0.9000
 
-* Support hydroscope's multi-page front end
+* Support hydroscope's multi-page front-end
+* Add station's and time series' data from all Hydroscope's domains
 
-# hydroscoper 0.1.0
+# hydroscoper 0.1.0 (Release date: 2017-12-22)
 
-* First release.
+* Initial submission to CRAN.